+++ packages/core/src/anyKeyed.test.ts
@@ -311,7 +311,7 @@ test("rejection of one of the sources should reject result", () => {
     a: new LazyPromise<"a">(() => () => {
       log("dispose a");
     }),
-    b: new LazyPromise((sink) => {
+    b: new LazyPromise<never>((sink) => {
       setTimeout(() => {
         sink.reject("oops");
       }, 1000);